tools/requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
fpdf2>=2.7.0
gspread>=6.0.0
google-auth>=2.0.0
